Split Unit and Integration tests
