+++ tests/unit/test_challenge.py
@@ -1,6 +1,6 @@
 import pytest
 from cryptography.hazmat.primitives.asymmetric.ed25519 import Ed25519PublicKey
-from tests.fixtures import *  # noqa: F403
+from tests.unit.fixtures import *  # noqa: F403
 from theoriq.agent import Agent, AgentConfig
